+++ package.json
@@ -18,6 +18,8 @@
   "devDependencies": {
     "@eslint/eslintrc": "^3",
     "@tailwindcss/postcss": "^4",
+    "@types/node": "22.15.14",
+    "@types/react": "19.1.3",
     "daisyui": "^5.0.35",
     "eslint": "^9",
     "eslint-config-next": "15.3.1",
