+++ src/app/globals.css
@@ -1,4 +1,17 @@
+@import url('https://fonts.googleapis.com/css2?family=Crimson+Pro:ital,wght@0,200..900;1,200..900&family=Playfair+Display:ital,wght@0,400..900;1,400..900&display=swap');
+
 @import "tailwindcss";
+
 @plugin "daisyui" {
   themes: light --default, night, fantasy;
+}
+
+h1,
+h2,
+header {
+  font-family: 'Playfair Display', serif;
+}
+
+button {
+  cursor: pointer;
 }
